requirements.txt
# Core dependencies for VedaCore KP Astrology Engine
# Synced with pyproject.toml - DO NOT MODIFY SEPARATELY

# Core astronomical calculations
pyswisseph==2.10.3.2
numpy==1.26.4  # Pinned for Numba compatibility
numba==0.59.1  # Latest stable for Python 3.11
scipy==1.14.0  # For advanced calculations
pandas==2.2.0  # Data processing

# Date/Time handling
pytz==2024.1
python-dateutil==2.9.0

# FastAPI and server - LATEST STABLE VERSIONS
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.0
python-multipart==0.0.12

# HTTP client for testing
httpx==0.27.0

# Monitoring
prometheus-client==0.21.0

# Testing - LATEST STABLE VERSIONS
pytest==8.3.0
pytest-cov==5.0.0
pytest-asyncio==0.24.0
pytest-mock==3.14.0
requests==2.32.3  # For API testing

# Performance
orjson==3.10.7  # Fast JSON serialization
aiofiles==24.1.0  # Async file operations

# Configuration
PyYAML==6.0.2

# Optional: Cloud storage for log rotation
boto3==1.35.0  # AWS S3
google-cloud-storage==2.18.0  # Google Cloud Storage

# Development tools (optional, install with pip install -r requirements-dev.txt)
# black==24.8.0
# ruff==0.6.0
# mypy==1.11.0
# pre-commit==3.8.0